requirements.txt
pyaudio
SpeechRecognition
--extra-index-url https://download.pytorch.org/whl/cu116
torch
realtimestt
pygame
openai
python_weather
icrawler
spotipy
pyttsx3
python-dotenv
git+https://github.com/openai/whisper.git